+++ RazorSvelte/App/index.svelte
@@ -2,6 +2,8 @@
     import Layout from "./shared/layout/default";
     import Modal from "./shared/components/modal.svelte";
     import Offcanvas from "./shared/components/offcanvas.svelte";
+    import ChartBox from "./shared/components/chart-box.svelte";
+    import urls from "./shared/urls";
 
     let selectionValue: string = "";
     let selectionElement: HTMLSelectElement;
@@ -69,6 +71,32 @@
             </div>
         </div>
 
+        <div class="shadow-lg card mt-3">
+            <div class="row">
+                <div class="col-md-4 chart">
+                    <ChartBox 
+                    type="bar" 
+                    getUrl={urls.chart1Url} 
+                    title="Number of employees by area - for the top 3 companies by the number of employees" />
+                </div>
+                <div class="col-md-4 chart">
+                    <ChartBox 
+                        type="line" 
+                        getUrl={urls.chart2Url} 
+                        title="Top 5 companies by the number of employees - employee growth last 10 years" />
+                </div>
+                <div class="col-md-4 hart">
+                    <ChartBox 
+                        type="pie" 
+                        getUrl={urls.chart3Url} 
+                        minHeight="300px"
+                        displayLegend={true}
+                        title="Top 10 companies by the country" />
+                </div>
+            </div>
+        </div>
+
+        <br />
     </div>
 </Layout>
 
@@ -105,3 +133,8 @@
     </div>
 </Offcanvas>
 
+<style lang="scss">
+    .chart {
+        padding: 25px;
+    }
+</style>